skill

+++ routes/web.php
@@ -1,6 +1,7 @@
 <?php
 
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -19,6 +20,8 @@ Auth::routes();
     Route::middleware(['auth'])->group(function () {
     //After Login the routes are accept by the loginUsers...
         Route::get('/home', [App\Http\Controllers\ProfileController::class,'index'])->name('admin.en.home');
+
+        Route::get('/buy/products', [App\Http\Controllers\ProductController::class,'index'])->name('admin.products');
 
         Route::get('/profile',[App\Http\Controllers\ProfileController::class,'show'])->name('admin.en.profile');
         Route::put('/profile/store/image',[App\Http\Controllers\ProfileController::class,'storeImage'])->name('store.profileImg');
@@ -43,7 +46,9 @@ Auth::routes();
         Route::post('/cart/clear/discount', [App\Http\Controllers\BasketController::class,'removeDiscount'])->name('remove.discount');
 
         Route::get('/checkout', [App\Http\Controllers\CheckoutController::class,'index'])->name('checkout.index');
-        Route::post('/checkout/store', [App\Http\Controllers\CheckoutController::class,'store'])->name('checkout.store');
+        Route::get('/checkout/index', [App\Http\Controllers\CheckoutController::class,'indexS'])->name('checkout.index.secret');
+
+        Route::post('/payment', [App\Http\Controllers\CheckoutController::class,'setPayment']);
 
         Route::get('/orders', [App\Http\Controllers\OrderController::class, 'index'])->name('order.index');
         Route::post('/order/items', [App\Http\Controllers\OrderController::class, 'store'])->name('order.store');
@@ -67,8 +72,11 @@ Auth::routes();
         })->name('error');
 
         Route::put('/password/update/{id}', [App\Http\Controllers\ProfileController::class,'update'])->name('password.dashboard.update');
+        Route::get('/search/employees', [App\Http\Controllers\UserController::class,'searchEmployees'])->name('user.search.employees');
+
 
     Route::middleware(['is_admin'])->group(function () {
+        Route::get('/search/all/employees', [App\Http\Controllers\UserController::class,'searchAllEmployees'])->name('user.search.all.employees');
         Route::get('/admin/users', [App\Http\Controllers\UserController::class,'index'])->name('users.index');
         Route::get('/admin/search/user', [App\Http\Controllers\UserController::class,'searchUser'])->name('user.search');
         Route::get('/admin/info/user/{id}', [App\Http\Controllers\UserController::class,'show'])->name('user.info');
@@ -76,6 +84,7 @@ Auth::routes();
         Route::get('/admin/edit/user/{id}', [App\Http\Controllers\UserController::class,'edit'])->name('user.edit');
         Route::put('/admin/update/user/{id}', [App\Http\Controllers\UserController::class,'update'])->name('user.update');
         Route::get('/search/user', [App\Http\Controllers\UserController::class,'searchUser'])->name('user.admin.search');
+        Route::post('/admin/existing/employee', [App\Http\Controllers\UserController::class,'addExistingEmployee'])->name('user.admin.exist');
 
         Route::get('/admin/orders', [App\Http\Controllers\OrderController::class,'allOrders'])->name('orders.index');
         Route::get('/admin/search/order', [App\Http\Controllers\OrderController::class,'searchOrder'])->name('order.search');
@@ -83,20 +92,29 @@ Auth::routes();
         Route::get('/admin/edit/order/{id}', [App\Http\Controllers\OrderController::class,'edit'])->name('order.edit');
         Route::put('/admin/update/order/{id}', [App\Http\Controllers\OrderController::class,'update'])->name('order.update');
 
-        Route::get('/admin/packages', [App\Http\Controllers\PackageController::class,'getAllPackages'])->name('packages.index');
-        Route::delete('/admin/packages/delete/{id}', [App\Http\Controllers\PackageController::class,'destroy'])->name('packages.delete');
+        Route::get('/admin/packages', [App\Http\Controllers\PackageController::class,'getAllPackages'])->name('packages.admin.index');
+        Route::delete('/admin/packages/delete/{id}', [App\Http\Controllers\PackageController::class,'destroy'])->name('packages.admin.delete');
         Route::post('/admin/add/packages', [App\Http\Controllers\PackageController::class,'store'])->name('packages.admin.add');
-        Route::get('/admin/search/packages', [App\Http\Controllers\PackageController::class,'searchPackage'])->name('package.search');
-        Route::get('/admin/edit/package/{id}', [App\Http\Controllers\PackageController::class,'edit'])->name('package.edit');
-        Route::get('/admin/edit/packageOwner/{id}', [App\Http\Controllers\PackageController::class,'editOwner'])->name('package.owner');
-        Route::put('/admin/update/package/{id}', [App\Http\Controllers\PackageController::class,'update'])->name('package.update');
-        Route::put('/admin/updateOwner/package/{id}', [App\Http\Controllers\PackageController::class,'updateOwner'])->name('package.updateOwner');
+        Route::get('/admin/search/packages', [App\Http\Controllers\PackageController::class,'searchPackage'])->name('packages.admin.search');
+        Route::get('/admin/edit/package/{id}', [App\Http\Controllers\PackageController::class,'edit'])->name('packages.admin.edit');
+        Route::get('/admin/edit/packageOwner/{id}', [App\Http\Controllers\PackageController::class,'editOwner'])->name('packages.admin.owner');
+        Route::put('/admin/update/package/{id}', [App\Http\Controllers\PackageController::class,'update'])->name('packages.admin.update');
+        Route::put('/admin/updateOwner/package/{id}', [App\Http\Controllers\PackageController::class,'updateOwner'])->name('packages.admin.updateOwner');
+
+        Route::get('/admin/certificates', [App\Http\Controllers\CertificateController::class,'getAllCertificates'])->name('certificates.admin.index');
+        Route::delete('/admin/certificate/delete/{id}', [App\Http\Controllers\CertificateController::class,'destroy'])->name('certificates.admin.delete');
+        Route::get('/admin/search/certificate', [App\Http\Controllers\CertificateController::class,'searchCertificate'])->name('certificates.admin.search');
+
         Route::get('/search', [App\Http\Controllers\UserController::class,'search'])->name('user.search');
+
+        Route::get('/add/product', [App\Http\Controllers\ProductController::class,'create'])->name('admin.create.product');
+        Route::post('/add/product', [App\Http\Controllers\ProductController::class,'store'])->name('admin.add.product');
     });
 });
 
         //Front end routes, User experience
         Route::get('/', function(){return view("front.landing");})->name('home');
+        Route::get('/info/about/admin', [App\Http\Controllers\ProductController::class, 'info'])->name('infoAdmin');
         Route::get('/products', function(){return view("front.products");})->name('front.products');
         Route::get('/team', function(){return view("front.teamTraining");})->name('front.team');
         Route::get('/contact', function(){return view("front.contact");})->name('front.contact');
@@ -105,6 +123,10 @@ Auth::routes();
         Route::get('/verify/certificate', function(){return view("front.verify");})->name('front.verify');
         Route::get('/cookies', function(){return view("front.cookies");})->name('front.cookies');
         Route::get('/product/{product}', [App\Http\Controllers\ProductController::class, 'show'])->name('front.product');
+        Route::get('/infoPdf', function(){
+            $path = "pdf/info.pdf";
+            return response()->download($path, 'info.pdf');
+        })->name('infoPdf');
 
 
         Route::get('/controller/payment/management',[App\Http\Controllers\EmployeeController::class,'controller'])->name('controller.payment.management');
